Fixing Benchmarking
